% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_conf_int.R
\name{stat_conf_int}
\alias{stat_conf_int}
\title{Create Confidence Intervals}
\usage{
stat_conf_int(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{conf_int supports the passing of additional
parameters, allowing for customization of graphics seen in other stats and
geoms. For more information of what additional arguments can be passed, look
at the help files for ggplot() and any stats and geoms being used.}

\item{aes(y)}{The only parameter required of this function is a y parameter
passed through the aes() function of the ggplot() call.}
}
\value{
This function in combination with a call to ggplot() and a call to
stat_per_sum returns a ggplot2 graphic object that can
display a percentile summary graph with
confidence intervals.
}
\description{
Add confidence intervals to EDF plots.
}
\details{
The function conf_int creates a custom stat for
{ggplot2} to add confidence intervals to other EDF stats.
}
\examples{
## Display percentile summary graph of mpg, with blue scatterplot and solid 
## quantile lines, and dashed confidence intervals
ggplot2::ggplot(mtcars, ggplot2::aes(y = mpg)) + 
stat_per_sum(color = "blue") +
stat_quant() +
stat_conf_int(linetype = "dashed")

}
\author{
Kathryn Cronquist

Thea De Jong

Ben Stano
}
