% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_quant.R
\name{stat_quant}
\alias{stat_quant}
\title{Add Quantile Lines to Percentile Summary Graphs}
\usage{
stat_quant(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{stat_per_sum and stat_quant support the passing of additional
parameters, allowing for customization of graphics seen in other stats and
geoms. For more information of what additional arguments can be passed, look
at the help files for ggplot() and any stats and geoms being used.}

\item{aes(y)}{The only parameter required of this function is a y parameter
passed through the aes() function of the ggplot() call.}
}
\value{
These functions in combination with a call to ggplot() return a
ggplot2 graphic object that can display a percentile summary graph.
}
\description{
Create lines displaying the 0\%, 25\%, 50\%, 75\%, and 100\% quantile
separations.
}
\details{
The functions stat_per_sum and stat_quant create a custom stat for
{ggplot2} to create a scatterplot of ({\frac{i}{n}},{X_i}) and an
accompanying stat for {ggplot2} that indicates the quantiles. Combined, these
graphics create a percentile summary graph to show the distribution of a
variable in a data set.
}
\examples{
## Display percentile summary graph of mpg, with blue scatterplot and dashed 
## quantile lines

ggplot2::ggplot(mtcars, ggplot2::aes(y = mpg)) + 
stat_per_sum(color = "blue") +
stat_quant(linetype = "dashed")

}
\author{
Kathryn Cronquist

Thea De Jong

Ben Stano
}
