% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cum_per_poly.R
\docType{data}
\name{StatCPP}
\alias{StatCPP}
\title{Create Cumulative Percentage Polygon Graphs}
\format{
An object of class \code{StatPercentile} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatCPP
}
\arguments{
\item{aes(x)}{The only parameter required of this function is an x parameter
passed through the aes() function. This parameter may be passed directly to
stat_cum_per_poly or to a previous ggplot() call.}

\item{...}{stat_cum_per_poly supports the passing of additional parameters,
allowing for customization of graphics seens in other stats and geoms. For
more information of what additional arguments can be passed, look at the
help files for ggplot() and any stats and geoms being used.}
}
\value{
This function in combination with a call to ggplot() returns a
ggplot2 graphic object that can display a cumulative percentage polygon.
}
\description{
Create graphic where data is displayed as a line plot
connecting the points (\eqn{X_i}, \eqn{\frac{i}{n}}).
}
\details{
This function creates a custom stat for {ggplot2} that can create
lines graphs showing the cumulative density polygon for a set of data. The
resulting graphic displays what percentage of the date is cumulatively
containing at a particular value. Such graphics give insight into the
distribution and shape of a data set.
}
\examples{
## Display CPP for the hwy variable in the mpg data set
ggplot(data = mpg, aes(x = hwy)) +
stat_cum_per_poly()

}
\author{
Ben Stano

Kathryn

Thea
}
\keyword{datasets}
